Completed appearence

+++ week7/assets/js/backoffice.js
@@ -3,18 +3,20 @@ const PRODUCTS_URL = "https://striveschool-api.herokuapp.com/api/product/";
 let prodId = new URLSearchParams(window.location.search).get("prodId");
 
 const backDel = async () => {
-  let response = await fetch(PRODUCTS_URL + prodId, {
-    method: "DELETE",
-    headers: {
-      Authorization:
-        "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDE0MzZkNWY4MWI0MjAwMTM5YjI4MTgiLCJpYXQiOjE2NzkwNDYzNTgsImV4cCI6MTY4MDI1NTk1OH0.EqBKFRGf984_UZ9sfQfkWVXYQ_oRZ3bReBpFutHxaIY",
-      "Content-Type": "application/json"
-    },
-  });
-  if (response.ok) {
-    alert("Hai eliminato il prodotto.");
-  } else {
-    alert("ERRORE: Non è stato possibile eliminare il prodotto");
+  if (confirm("Sei sicuro di voler eliminare il prodotto?")) {
+    let response = await fetch(PRODUCTS_URL + prodId, {
+      method: "DELETE",
+      headers: {
+        Authorization:
+          "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDE0MzZkNWY4MWI0MjAwMTM5YjI4MTgiLCJpYXQiOjE2NzkwNDYzNTgsImV4cCI6MTY4MDI1NTk1OH0.EqBKFRGf984_UZ9sfQfkWVXYQ_oRZ3bReBpFutHxaIY",
+        "Content-Type": "application/json"
+      },
+    });
+    if (response.ok) {
+      alert("Hai eliminato il prodotto.");
+    } else {
+      alert("ERRORE: Non è stato possibile eliminare il prodotto");
+    }
   }
 };
 
@@ -89,6 +91,18 @@ formReference.addEventListener("submit", (e) => {
   };
   queryProd(newProd);
 });
+
+const resetForm = () => {
+  if (confirm("Sei sicuro di voler cancellare tutti i dati inseriti nel form?")) {
+    document.getElementById("prodName").value = "";
+    document.getElementById("prodDesc").value = "";
+    document.getElementById("prodBrand").value = "";
+    document.getElementById("prodImg").value = "";
+    document.getElementById("prodPrice").value = "";
+  }
+};
+
+document.getElementById("reset").addEventListener("click", resetForm);
 
 
 /* AUTHORIZATION :
